+++ core/src/main/scala/kafka/common/MessageReader.scala
@@ -29,7 +29,7 @@ import org.apache.kafka.clients.producer.ProducerRecord
   * This is used by the `ConsoleProducer`.
   */
 @deprecated("This class has been deprecated and will be removed in 4.0." +
-  "Please use org.apache.kafka.tools.RecordReader instead", "3.5.0")
+  "Please use org.apache.kafka.tools.api.RecordReader instead", "3.5.0")
 trait MessageReader {
 
   def init(inputStream: InputStream, props: Properties): Unit = {}
