+++ settings.gradle
@@ -58,6 +58,7 @@ include 'clients',
     'streams:upgrade-system-tests-32',
     'streams:upgrade-system-tests-33',
     'tools',
+    'tools:tools-api',
     'trogdor'
 
 rootProject.name = 'kafka'
