+++ tools/tools-api/src/test/java/org/apache/kafka/tools/api/RecordReaderTest.java
@@ -1,0 +1,33 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements. See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.apache.kafka.tools.api;
+
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.Test;
+
+import java.util.Collections;
+
+public class RecordReaderTest {
+
+    @Test
+    void testDefaultCloseAndConfigure() {
+        RecordReader reader = inputStream -> null;
+        // `configure` and `close` should have default empty body
+        Assertions.assertDoesNotThrow(() -> reader.configure(Collections.emptyMap()));
+        Assertions.assertDoesNotThrow(reader::close);
+    }
+}
